backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
Flask-Mail==0.10.0
Werkzeug==3.0.3
SQLAlchemy==2.0.31
redis==5.0.7
celery==5.4.0
celery[redis]==5.4.0
flower==2.0.1
python-dotenv==1.0.1
requests==2.32.3
Jinja2==3.1.4
